Attempted to move methods from gameengine to grid:
program not working, will try to fix it by tonight
