+++ src/edu/csupomona/cs/cs141/classproject/GameEngine.java
@@ -15,105 +15,116 @@ public class GameEngine {
 	//and the movement of the ninjas.
 
 	private GridMember[][] grid;
+	
+	public Grid gridm;
 
 	private Ninja[] ninjas;
 
-	private Taha thePlayer;
+	private Taha tahaPlayer;
 
 	private GridMember player;
 
 	private int[] playerPosition = new int[2];
-
-	public GameEngine(Taha tahaPlayer) {
-
-		ninjas = new Ninja[6];
-		for (int x = 0; x < ninjas.length; x++) {
-			ninjas[x] = new Ninja();
-		}
-		grid = new GridMember[9][9];
-		player = tahaPlayer;
-		thePlayer = tahaPlayer;
-		player = thePlayer;
-
-		for (int k = 0; k < grid.length; k++) {
-			for (int l = 0; l < grid[k].length; l++) {
-				grid[k][l] = new EmptyMember();
-			}
-		}
-
-		spawnRooms();
-
-		// spawns ninjas randomly, I am having trouble with having them spawn
-		// everywhere, not within
-		// 3 spaces of the player.
-		spawnNinjas();
-
-		//ninja = new Ninja();
-		//grid[5][5] = ninja;
-		//ninja.setPosition(5, 5);
-		grid[8][0] = player;
-		thePlayer.setPosition(8, 0);
-
-		playerPosition[0] = 8;
-		playerPosition[1] = 0;
+	
+	private int playerDirection;
+	
+	public GameEngine(){
+		gridm = new Grid(tahaPlayer);
 	}
 
-	public void spawnRooms() {
-		Random rand = new Random();
-		int randRoom = rand.nextInt(9);
-		int startRow = 1;
-		int startCol = 1;
-		int count = 0;
-		for (int y = 0; y < 3; y++) {
+// Will be removed soon... (moved to grid) -Taha
+//	public GameEngine(Taha tahaPlayer) {
+//
+//		ninjas = new Ninja[6];
+//		for (int x = 0; x < ninjas.length; x++) {
+//			ninjas[x] = new Ninja();
+//		}
+//		grid = new GridMember[9][9];
+//		player = tahaPlayer;
+//		thePlayer = tahaPlayer;
+//		player = thePlayer;
+//
+//		for (int k = 0; k < grid.length; k++) {
+//			for (int l = 0; l < grid[k].length; l++) {
+//				grid[k][l] = new EmptyMember();
+//			}
+//		}
+//
+//		spawnRooms();
+//
+//		// spawns ninjas randomly, I am having trouble with having them spawn
+//		// everywhere, not within
+//		// 3 spaces of the player.
+//		spawnNinjas();
+//
+//		//ninja = new Ninja();
+//		//grid[5][5] = ninja;
+//		//ninja.setPosition(5, 5);
+//		grid[8][0] = player;
+//		thePlayer.setPosition(8, 0);
+//
+//		playerPosition[0] = 8;
+//		playerPosition[1] = 0;
+//	}
+//
+//	public void spawnRooms() {
+//		Random rand = new Random();
+//		int randRoom = rand.nextInt(9);
+//		int startRow = 1;
+//		int startCol = 1;
+//		int count = 0;
+//		for (int y = 0; y < 3; y++) {
+//
+//			Room room1 = new Room();
+//			grid[startRow][startCol] = room1;
+//			++count;
+//			if (randRoom == count) {
+//				room1.giveBriefcase();
+//			}
+//
+//			Room room2 = new Room();
+//			grid[startRow][startCol + 3] = room2;
+//			++count;
+//			if (randRoom == count) {
+//				room2.giveBriefcase();
+//			}
+//
+//			Room room3 = new Room();
+//			grid[startRow][startCol + 6] = room3;
+//			++count;
+//			if (randRoom == count) {
+//				room3.giveBriefcase();
+//			}
+//
+//			startRow += 3;
+//		}
+//	}
+//
+//	public void spawnNinjas() {
+//		Random rand = new Random();
+//		int[] ninjaCoordinates = new int[2];
+//		for (Ninja currentNinja : ninjas) {
+//			int ninRow = rand.nextInt(6) + 3;
+//			int ninCol = rand.nextInt(6) + 3;
+//			while (grid[ninRow][ninCol].isSomething()) {
+//				ninRow = rand.nextInt(6) + 3;
+//				ninCol = rand.nextInt(6) + 3;
+//			}
+//			ninjaCoordinates[0] = ninRow;
+//			ninjaCoordinates[1] = ninCol;
+//
+//			grid[ninRow][ninCol] = currentNinja;
+//			currentNinja.setPosition(ninRow, ninCol);
+//		}
+//
+//		grid[8][0] = player;
+//		thePlayer.setPosition(8, 0);
+//
+//		playerPosition[0] = 8;
+//		playerPosition[1] = 0;
+//	}
 
-			Room room1 = new Room();
-			grid[startRow][startCol] = room1;
-			++count;
-			if (randRoom == count) {
-				room1.giveBriefcase();
-			}
 
-			Room room2 = new Room();
-			grid[startRow][startCol + 3] = room2;
-			++count;
-			if (randRoom == count) {
-				room2.giveBriefcase();
-			}
-
-			Room room3 = new Room();
-			grid[startRow][startCol + 6] = room3;
-			++count;
-			if (randRoom == count) {
-				room3.giveBriefcase();
-			}
-
-			startRow += 3;
-		}
-	}
-
-	public void spawnNinjas() {
-		Random rand = new Random();
-		int[] ninjaCoordinates = new int[2];
-		for (Ninja currentNinja : ninjas) {
-			int ninRow = rand.nextInt(6) + 3;
-			int ninCol = rand.nextInt(6) + 3;
-			while (grid[ninRow][ninCol].isSomething()) {
-				ninRow = rand.nextInt(6) + 3;
-				ninCol = rand.nextInt(6) + 3;
-			}
-			ninjaCoordinates[0] = ninRow;
-			ninjaCoordinates[1] = ninCol;
-
-			grid[ninRow][ninCol] = currentNinja;
-			currentNinja.setPosition(ninRow, ninCol);
-		}
-
-		grid[8][0] = player;
-		thePlayer.setPosition(8, 0);
-
-		playerPosition[0] = 8;
-		playerPosition[1] = 0;
-	}
 
 	public void printGrid() { // this will eventually be moved to UserInterface
 
@@ -126,15 +137,42 @@ public class GameEngine {
 		System.out.println();
 
 	}
+// Will be removed soon... (moved to grid) -Taha
+//	public Taha getPlayer() {
+//		return thePlayer;
+//	}
 
-	public Taha getPlayer() {
-		return thePlayer;
-	}
+//	public void move(String direction, Grid grid){ //going to move this to the GameEngine
+//	int[] playerPosition = grid.getPlayerPostion();
+//	int row = playerPosition[0];
+//	int col = playerPosition[1];
+//	switch(direction){
+//	
+//	case "w":
+//		playerDirection = 1;
+//		movePlayer(row-1, col);
+//		break;
+//	case "d":
+//		playerDirection = 2;
+//		movePlayer(row, col+1);
+//		break;
+//	case "s":
+//		playerDirection = 3;
+//		movePlayer(row+1, col);
+//		break;
+//	case "a":
+//		playerDirection = 4;
+//		movePlayer(row, col-1);
+//		break;
+//	}
+//	
+//	
+//}
 
 	public void movePlayer(int row, int col) {
 		int previousRow = playerPosition[0];
 		int previousCol = playerPosition[1];
-		int playerDirection = thePlayer.getPlayerDirection();
+		int playerDirection = tahaPlayer.getPlayerDirection();
 
 		// this try-catch block deals with moving the player object to a new
 		// position on the Grid.
@@ -165,7 +203,7 @@ public class GameEngine {
 				playerPosition[0] = row;
 				playerPosition[1] = col;
 				grid[row][col] = player;
-				thePlayer.setPosition(row, col);
+				tahaPlayer.setPosition(row, col);
 				grid[previousRow][previousCol] = new EmptyMember();
 
 			    moveNinjas(previousRow, previousCol);
@@ -182,6 +220,7 @@ public class GameEngine {
 		}
 
 	}
+	
 	
 	public void resetPlayerSeeAround(){
 		int row = playerPosition[0];
@@ -323,16 +362,16 @@ public class GameEngine {
 		}
 	}
 
-
-	public GridMember getGridMember(int row, int col) {
-
-		GridMember gridMember = grid[row][col];
-		return gridMember;
-
-	}
-
-	public int[] getPlayerPostion() {
-		return playerPosition;
-	}
+// Will be removed soon... (moved to grid) -Taha
+//	public GridMember getGridMember(int row, int col) {
+//
+//		GridMember gridMember = grid[row][col];
+//		return gridMember;
+//
+//	}
+//
+//	public int[] getPlayerPostion() {
+//		return playerPosition;
+//	}
 
 }
